Move parsing of "cargo metadata" output from Cargo to CargoProjectImpl
Cargo now only returns a JSON String.
