+++ src/main/kotlin/org/rust/cargo/project/model/impl/CargoProjectImpl.kt
@@ -5,6 +5,8 @@
 
 package org.rust.cargo.project.model.impl
 
+import com.google.gson.Gson
+import com.google.gson.JsonSyntaxException
 import com.intellij.execution.ExecutionException
 import com.intellij.execution.process.ProcessAdapter
 import com.intellij.execution.process.ProcessEvent
@@ -50,6 +52,7 @@ import org.rust.cargo.project.workspace.StandardLibrary
 import org.rust.cargo.runconfig.command.workingDirectory
 import org.rust.cargo.toolchain.RustToolchain
 import org.rust.cargo.toolchain.Rustup
+import org.rust.cargo.toolchain.impl.CargoMetadata
 import org.rust.cargo.util.AutoInjectedCrates
 import org.rust.cargo.util.DownloadResult
 import org.rust.ide.notifications.showBalloon
@@ -499,7 +502,7 @@ private fun fetchCargoWorkspace(
         }
         val cargo = toolchain.cargoOrWrapper(projectDirectory)
         try {
-            val ws = cargo.fullProjectDescription(project, projectDirectory, object : ProcessAdapter() {
+            val json = cargo.fullProjectDescription(project, projectDirectory, object : ProcessAdapter() {
                 override fun onTextAvailable(event: ProcessEvent, outputType: Key<Any>) {
                     val text = event.text.trim { it <= ' ' }
                     if (text.startsWith("Updating") || text.startsWith("Downloading")) {
@@ -507,6 +510,15 @@ private fun fetchCargoWorkspace(
                     }
                 }
             })
+
+            val rawData = try {
+                Gson().fromJson(json, CargoMetadata.Project::class.java)
+            } catch (e: JsonSyntaxException) {
+                throw ExecutionException(e)
+            }
+            val projectDescriptionData = CargoMetadata.clean(rawData)
+            val manifestPath = projectDirectory.resolve("Cargo.toml")
+            val ws = CargoWorkspace.deserialize(manifestPath, projectDescriptionData)
             ok(ws)
         } catch (e: ExecutionException) {
             err(e.message ?: "failed to run Cargo")
